[Delivers #71702378] Student prices are now shown in list and details
