+++ build.gradle
@@ -1,6 +1,6 @@
 def versionMajor = 1 // incompatible API changes
-def versionMinor = 6 // added functionality in a backwards-compatible manner
-def versionPatch = 1 // backwards-compatible bug fixes
+def versionMinor = 7 // added functionality in a backwards-compatible manner
+def versionPatch = 0 // backwards-compatible bug fixes
 def versionBuild = 0 // bump for dogfood builds, public betas, etc.
 
 buildscript {
